refactor: move db queries + drizzle schema to @dtx/common/server
Leaves only the SvelteKit-coupled getDb(platform) wrapper in
dtx-web/src/lib/server/db.ts. Phase 0 of the API server migration.
