+++ packages/dtx-web/src/routes/api/chart/+server.ts
@@ -1,8 +1,15 @@
 import { json } from '@sveltejs/kit';
-import { getDb, listSimfiles, createSimfile, createDtxFiles, deleteSimfile } from '$lib/server/db';
-import type { SimfileWithDtxFiles } from '$lib/server/db';
+import { getDb } from '$lib/server/db';
 import { toSimfileWithDtx } from '@dtx/common';
-import { isPreviewKey, logger } from '@dtx/common/server';
+import {
+	isPreviewKey,
+	logger,
+	listSimfiles,
+	createSimfile,
+	createDtxFiles,
+	deleteSimfile,
+	type SimfileWithDtxFiles
+} from '@dtx/common/server';
 
 const MAX_CONCURRENT_R2_CHECKS = 4;
