+++ packages/dtx-web/src/routes/api/chart/next-display-id/+server.ts
@@ -1,6 +1,6 @@
 import { json } from '@sveltejs/kit';
-import { getDb, getNextDisplayId } from '$lib/server/db';
-import { logger } from '@dtx/common/server';
+import { getDb } from '$lib/server/db';
+import { logger, getNextDisplayId } from '@dtx/common/server';
 
 export const GET = async ({ platform, locals }: { platform: App.Platform; locals: App.Locals }) => {
 	const user = locals.user;
