+++ packages/dtx-web/src/routes/api/chart/search/+server.ts
@@ -1,6 +1,6 @@
 import { json } from '@sveltejs/kit';
-import { getDb, searchSimfiles } from '$lib/server/db';
-import { logger } from '@dtx/common/server';
+import { getDb } from '$lib/server/db';
+import { logger, searchSimfiles } from '@dtx/common/server';
 
 /** GET /api/chart/search — Search charts by title/artist */
 export const GET = async ({
