+++ packages/dtx-web/src/routes/api/simFile/delete/[simfileID]/server.test.ts
@@ -2,7 +2,8 @@ import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { DELETE } from './+server';
 import type { R2Bucket } from '@cloudflare/workers-types';
 import type { Session } from '@supabase/supabase-js';
-import { getDb, getSimfileOwner, deleteSimfile } from '$lib/server/db';
+import { getDb } from '$lib/server/db';
+import { getSimfileOwner, deleteSimfile } from '@dtx/common/server';
 
 // Mock logger
 vi.mock('@dtx/common/server', () => ({
@@ -10,10 +11,14 @@ vi.mock('@dtx/common/server', () => ({
 		error: vi.fn(),
 		info: vi.fn(),
 		warn: vi.fn()
-	}
+	},
+	getSimfileOwner: vi.fn(),
+	deleteSimfile: vi.fn()
 }));
 
-vi.mock('$lib/server/db');
+vi.mock('$lib/server/db', () => ({
+	getDb: vi.fn()
+}));
 
 // Create a mock Request object
 const createMockRequest = (method: string, url: string, headers?: Headers): Request => {
